backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
anthropic
pymupdf
python-docx
httpx
beautifulsoup4
python-dotenv
